[FEATURE] ViewHelper to render srcset attribute
This change introduces a new SrcsetViewHelper, which renders
the HTML "srcset" attribute that is used for responsive images.
As the ViewHelper doesn't render a whole HTML tag, it can be used
in different locations:

* with custom `<img>` tags not using the ImageViewHelper
* with `<source>` tags inside of `<picture>`
* with `<link rel="preload" as="image" imagesrcset="..." />` to preload
  responsive images
* with the existing ImageViewHelper using `srcset="..."`

The change also adds a separate class structures to parse, calculate,
and render srcset attributes, independent of both Fluid and FAL.
This makes these classes quite flexible to be used in other locations
of the core, as well as in 3rd party extensions.

The ViewHelper's API tries to be simple and descriptive, which
is why it only allows FAL objects as input. This makes the underlying
processing much more intuitive to the user, as some of the old
scaling options (such as "c" and "m" modifiers) don't really make
sense with responsive images in mind.

Resolves: #102215
Releases: main
Change-Id: Ib050f74267b91e384ccfbb37dc7c07c99ed948a6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81494
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Ayke Halder <mail@ayke-halder.de>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Ayke Halder <mail@ayke-halder.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Garvin Hicking <garvin@hick.ing>